refactor: Breaks up main.py
Breaks up functions from main.py, with two new files. get_and_order is
now used for retrieval, parsing, and ordering functions. main_delete now
contains the main deletion functions (delete_resource and
retry_failed_deletions). Now main.py only contains main.
